✨ feat(packaging): move study runner into the package; add silisocs-study CLI
- study orchestration moves to silisocs.studies.run_study /
  study_artifacts; thin shims keep experiments/run_study.py and
  experiments/_internal/study_artifacts.py working for SLURM templates
  and existing scripts (re-export every public symbol)
- new silisocs-study console entry point
- study child runs now launch with the running interpreter (sys.executable)
  instead of always 'uv run python', which broke pip-installed venvs where
  uv resolves a different interpreter without silisocs; RUN_STUDY_PYTHON
  still overrides
- study_template_v1 bundled into the wheel; silisocs.studies.templates
  locates it for repo or pip installs
- docs/study_schema.md + experiments.md point to the new runner location

Verified in a clean pip venv from a neutral cwd: silisocs-study runs a
study referencing a bundled scenario by name (success), and a second
invocation skips the completed run (idempotent resume).
