+++ src/silisocs/studies/__init__.py
@@ -1,0 +1,1 @@
+"""Study orchestration: plan and execute study.yaml experiment grids."""
